- Optimized types in sendRC5ext and sendSharpAlt. - Added `DECODE_NEC_STANDARD` and `SEND_NEC_STANDARD`. - Renamed all IRrecv* examples to IRreceive*. - Added functions `printResultShort(&Serial)` and `getProtocolString(decode_type_t aDecodeType)`. - Added flag `results.isRepeat`. - Updated examples.

+++ src/ir_RC5_RC6.cpp
@@ -83,19 +83,19 @@ void IRsend::sendRC5(unsigned long data, int nbits) {
     space(0);  // Always end with the LED off
 }
 
-void IRsend::sendRC5ext(unsigned long addr, unsigned long cmd, boolean toggle) {
+void IRsend::sendRC5ext(uint8_t addr, uint8_t cmd, boolean toggle) {
     // Set IR carrier frequency
     enableIROut(36);
 
-    unsigned long addressBits = 5;
-    unsigned long commandBits = 7;
+    uint8_t addressBits = 5;
+    uint8_t commandBits = 7;
 //    unsigned long nbits = addressBits + commandBits;
 
     // Start
     mark(RC5_T1);
 
     // Bit #6 of the command part, but inverted!
-    unsigned long cmdBit6 = (1UL << (commandBits - 1)) & cmd;
+    uint8_t cmdBit6 = (1UL << (commandBits - 1)) & cmd;
     if (cmdBit6) {
         // Inverted (1 -> 0 = mark-to-space)
         mark(RC5_T1);
@@ -124,7 +124,7 @@ void IRsend::sendRC5ext(unsigned long addr, unsigned long cmd, boolean toggle) {
     }
 
     // Address
-    for (unsigned long mask = 1UL << (addressBits - 1); mask; mask >>= 1) {
+    for (uint8_t mask = 1UL << (addressBits - 1); mask; mask >>= 1) {
         if (addr & mask) {
             space(RC5_T1); // 1 is space, then mark
             mark(RC5_T1);
@@ -135,7 +135,7 @@ void IRsend::sendRC5ext(unsigned long addr, unsigned long cmd, boolean toggle) {
     }
 
     // Command
-    for (unsigned long mask = 1UL << (commandBits - 1); mask; mask >>= 1) {
+    for (uint8_t mask = 1UL << (commandBits - 1); mask; mask >>= 1) {
         if (cmd & mask) {
             space(RC5_T1); // 1 is space, then mark
             mark(RC5_T1);
